New "Event Story" section in 冒険の記録/Scene Select with all CB/BC stories
Subcategory names for named events match their original banner pictures
or (for untranslated banners) the "Clash Batte! Foo: The End" scenario text.
The name of the boss as shown in Clash Search is added, plus any other species
that isn't its pre-evolution but also appeared as a boss during the event.
For the Halloween event, which isn't present in Clash Search,
emulate the same pattern and name the fully evolved boss (Fantomon).

The order of subcategories matches the order
in which the scenarios were first shown in Japan.